Implement purchase flow and exception handling
Added new panels and logic for managing the purchase flow, including payment and receipt views. Introduced custom exceptions for card validation and expiration dates. Updated CarritoDAO to support product addition and removal. Refactored controller and view classes to support new actions and UI changes for cart and product management.

+++ proyecto/src/co/edu/unbosque/view/ProductoInfoPanel.java
@@ -63,7 +63,7 @@ public class ProductoInfoPanel extends JPanel {
         this.add(mercadoLibre);
     }
 
-    public void mostrarProductoInfo(float precio, String nombre, String descripcion, int unidades, String url, String atributo1, String atributo2) {
+    public void mostrarProductoInfo(double precio, String nombre, String descripcion, int unidades, String url, String atributo1, String atributo2) {
     	cajaFondoProducto.removeAll();
 
         ImageIcon icon = new ImageIcon(url);
